feat(MiniApp(User)): Make user mini-app instance
