+++ src/interfaces/http/router.js
@@ -47,7 +47,7 @@ module.exports = ({ config, logger, database }) => {
 
   apiRouter.use('/', controller('index'))
   apiRouter.use('/token', controller('token'))
-  apiRouter.use('/users', controller('user'))
+  apiRouter.use('/users', controller('user').router)
   apiRouter.use('/companies', controller('company'))
 
   router.use(`/api/${config.version}`, apiRouter)
